Pipes.Csv

Description

This module allows constant-space CSV parsing.

It feeds ByteStrings into cassavas incremental CSV parser to attain true constant-space record streaming.

Example

Heres a simple example that reads from stdin and writes to a file

    import Pipes.Safe (runSafeT)
    import qualified Pipes.Safe.Prelude  as PS
    import qualified Pipes.ByteString    as PB
    import Data.Vector (fromList)
    import System.IO (IOMode(WriteMode))

    data Person = Person String Int
                deriving (Show)

    instance FromNamedRecord Person where
      parseNamedRecord p =
        Person <$> p .: "name"
               <*> p .: "age"

    personRec ~(Person name age) = ["name" .= name, "age" .= age]

    instance ToNamedRecord Person where
      toNamedRecord = namedRecord . personRec

    persons :: Monad m => Producer ByteString m () -> Producer Person m ()
    persons p = decodeByName p >-> right

    -- note: right can be replaced with Pipes.Prelude.concat in ghc-7.8,
    --       thanks to a Foldable instance for Either
    right :: (Monad m) => Pipe (Either a b) b m r
    right = loop
      where
        loop = await >>= s -> case s of
          Left _  -> loop
          Right v -> yield v >> loop

    write f = PS.withFile f WriteMode PB.toHandle

    main = runSafeT $ runEffect $ pipeline
      where
        header = fromList $ map fst $ personRec undefined
        pipeline = persons stdin
               >-> right
               >-> encodeByName header
               >-> write "persons_out.csv"

Decode records

decode :: forall (m :: Type -> Type) a. (Monad m, FromRecord a) => HasHeader -> Producer ByteString m () -> Producer (Either String a) m ()

Equivalent to decodeWith defaultDecodeOptions.

decodeWith :: forall (m :: Type -> Type) a. (Monad m, FromRecord a) => DecodeOptions -> HasHeader -> Producer ByteString m () -> Producer (Either String a) m ()

Create a Producer that takes a ByteString Producer as input, producing either errors or FromRecords.

Decode named records

decodeByName :: forall (m :: Type -> Type) a. (Monad m, FromNamedRecord a) => Producer ByteString m () -> Producer (Either String a) m ()

Equivalent to decodeByNameWith defaultDecodeOptions.

decodeByNameWith :: forall (m :: Type -> Type) a. (Monad m, FromNamedRecord a) => DecodeOptions -> Producer ByteString m () -> Producer (Either String a) m ()

Create a Producer that takes a ByteString Producer as input, producing either errors or FromNamedRecords.

Decode parsed records

feedParser :: forall (m :: Type -> Type) a. Monad m => Parser a -> Producer ByteString m () -> Producer (Either String a) m ()

Create a Record Producer by feeding ByteStrings into a Parser

feedHeaderParser :: forall (m :: Type -> Type) a. Monad m => HeaderParser (Parser a) -> Producer ByteString m () -> Producer (Either String a) m ()

Create a NamedRecord Producer by feeding ByteStrings into a Parser

Encode records

encode :: forall (m :: Type -> Type) a r. (Monad m, ToRecord a) => Pipe a ByteString m r

Encode records as strict ByteStrings

encodeWith :: forall (m :: Type -> Type) a r. (Monad m, ToRecord a) => EncodeOptions -> Pipe a ByteString m r

Encode records as strict ByteStrings

Encode named records

encodeByName :: forall (m :: Type -> Type) a r. (Monad m, ToNamedRecord a) => Header -> Pipe a ByteString m r

Encode named records as strict ByteStrings

encodeByNameWith :: forall (m :: Type -> Type) a r. (Monad m, ToNamedRecord a) => EncodeOptions -> Header -> Pipe a ByteString m r

Encode named records as strict ByteStrings

Re-exports

Data.Csv re-exports common types and operators:

FromRecord, FromNamedRecord, ToRecord, ToNamedRecord

ToField, FromField

Record, Field, NamedRecord

(.!), (.:), (.=)

DecodeOptions, defaultDecodeOptions

type Field = ByteString

A single field within a record.

type NamedRecord = HashMap ByteString ByteString

A record corresponds to a single line in a CSV file, indexed by the column name rather than the column index.

type Header = Vector Name

The header corresponds to the first line a CSV file. Not all CSV files have a header.

type Record = Vector Field

A record corresponds to a single line in a CSV file.

data DecodeOptions

Options that controls how data is decoded. These options can be used to e.g. decode tab-separated data instead of comma-separated data.

To avoid having your program stop compiling when new fields are added to DecodeOptions, create option records by overriding values in defaultDecodeOptions. Example:

myOptions = defaultDecodeOptions {
      decDelimiter = fromIntegral (ord '\t')
    }

defaultDecodeOptions :: DecodeOptions

Decoding options for parsing CSV files.

class ToField a where

A type that can be converted to a single CSV field.

Example type and instance:

{-# LANGUAGE OverloadedStrings #-}

data Color = Red | Green | Blue

instance ToField Color where
    toField Red   = "R"
    toField Green = "G"
    toField Blue  = "B"

Methods

toField :: a -> Field

class FromField a where

A type that can be converted from a single CSV field, with the possibility of failure.

When writing an instance, use empty, mzero, or fail to make a conversion fail, e.g. if a Field can't be converted to the given type.

Example type and instance:

{-# LANGUAGE OverloadedStrings #-}

data Color = Red | Green | Blue

instance FromField Color where
    parseField s
        | s == "R"  = pure Red
        | s == "G"  = pure Green
        | s == "B"  = pure Blue
        | otherwise = mzero

Methods

parseField :: Field -> Parser a

class ToNamedRecord a where

A type that can be converted to a single CSV record.

An example type and instance:

data Person = Person { name :: !Text, age :: !Int }

instance ToNamedRecord Person where
    toNamedRecord (Person name age) = namedRecord [
        "name" .= name, "age" .= age]

Minimal complete definition

Nothing

Methods

toNamedRecord :: a -> NamedRecord

Convert a value to a named record.

default toNamedRecord :: (Generic a, GToRecord (Rep a) (ByteString, ByteString)) => a -> NamedRecord

class FromNamedRecord a where

A type that can be converted from a single CSV record, with the possibility of failure.

When writing an instance, use empty, mzero, or fail to make a conversion fail, e.g. if a Record has the wrong number of columns.

Given this example data:

name,age
John,56
Jane,55

here's an example type and instance:

{-# LANGUAGE OverloadedStrings #-}

data Person = Person { name :: !Text, age :: !Int }

instance FromNamedRecord Person where
    parseNamedRecord m = Person <$>
                         m .: "name" <*>
                         m .: "age"

Note the use of the OverloadedStrings language extension which enables ByteString values to be written as string literals.

Minimal complete definition

Nothing

Methods

parseNamedRecord :: NamedRecord -> Parser a

default parseNamedRecord :: (Generic a, GFromNamedRecord (Rep a)) => NamedRecord -> Parser a

class ToRecord a where

A type that can be converted to a single CSV record.

An example type and instance:

data Person = Person { name :: !Text, age :: !Int }

instance ToRecord Person where
    toRecord (Person name age) = record [
        toField name, toField age]

Outputs data on this form:

John,56
Jane,55

Minimal complete definition

Nothing

Methods

toRecord :: a -> Record

Convert a value to a record.

default toRecord :: (Generic a, GToRecord (Rep a) Field) => a -> Record

class FromRecord a where

A type that can be converted from a single CSV record, with the possibility of failure.

When writing an instance, use empty, mzero, or fail to make a conversion fail, e.g. if a Record has the wrong number of columns.

Given this example data:

John,56
Jane,55

here's an example type and instance:

data Person = Person { name :: !Text, age :: !Int }

instance FromRecord Person where
    parseRecord v
        | length v == 2 = Person <$>
                          v .! 0 <*>
                          v .! 1
        | otherwise     = mzero

Minimal complete definition

Nothing

Methods

parseRecord :: Record -> Parser a

default parseRecord :: (Generic a, GFromRecord (Rep a)) => Record -> Parser a

(.!) :: FromField a => Record -> Int -> Parser a infixl 9

Alias for index.

(.:) :: FromField a => NamedRecord -> ByteString -> Parser a

Alias for lookup.

(.=) :: ToField a => ByteString -> a -> (ByteString, ByteString)

Alias for namedField.

data EncodeOptions

Options that controls how data is encoded. These options can be used to e.g. encode data in a tab-separated format instead of in a comma-separated format.

To avoid having your program stop compiling when new fields are added to EncodeOptions, create option records by overriding values in defaultEncodeOptions. Example:

myOptions = defaultEncodeOptions {
      encDelimiter = fromIntegral (ord '\t')
    }

N.B. The encDelimiter must not be the quote character (i.e. ") or one of the record separator characters (i.e. \n or \r).

defaultEncodeOptions :: EncodeOptions

Encoding options for CSV files.

data HasHeader

Is the CSV data preceded by a header?

Constructors

HasHeader
The CSV data is preceded by a header

NoHeader
The CSV data is not preceded by a header